# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blockchain Freelance Marketplace API",
    "version": "1.0.0",
    "description": "A decentralized freelance marketplace with AI skill matching and blockchain payments",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "BlockchainRating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractId": {
            "type": "string"
          },
          "raterId": {
            "type": "string"
          },
          "rateeId": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer"
          },
          "transactionHash": {
            "type": "string"
          }
        }
      },
      "ReputationScore": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "description": "Weighted average score with time decay"
          },
          "totalRatings": {
            "type": "integer"
          },
          "averageRating": {
            "type": "number",
            "description": "Simple average without time decay"
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockchainRating"
            }
          }
        }
      },
      "WorkHistoryEntry": {
        "type": "object",
        "properties": {
          "contractId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "projectTitle": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "freelancer",
              "employer"
            ]
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "ratingComment": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/reputation/can-rate": {
      "get": {
        "summary": "Check if user can rate",
        "description": "Check if the authenticated user can rate another user for a specific contract",
        "tags": [
          "Reputation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Contract ID"
          },
          {
            "in": "query",
            "name": "rateeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID to rate"
          }
        ],
        "responses": {
          "200": {
            "description": "Check completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "canRate": {
                      "type": "boolean"
                    },
                    "reason": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/reputation/rate": {
      "post": {
        "summary": "Submit rating",
        "description": "Submit a rating for another user after contract completion. Rating must be between 1 and 5.",
        "tags": [
          "Reputation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Rating submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rating": {
                      "$ref": "#/components/schemas/BlockchainRating"
                    },
                    "transactionHash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid rating or validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Contract not found"
          },
          "409": {
            "description": "Duplicate rating"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "contractId",
                  "rateeId",
                  "rating"
                ],
                "properties": {
                  "contractId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "rateeId": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,35}$"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reputation/{userId}": {
      "get": {
        "summary": "Get user reputation",
        "description": "Retrieves the reputation score and ratings for a user from the blockchain",
        "tags": [
          "Reputation"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "User ID to get reputation for (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Reputation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationScore"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/reputation/{userId}/history": {
      "get": {
        "summary": "Get work history",
        "description": "Retrieves the work history for a user including completed contracts and ratings",
        "tags": [
          "Reputation"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "User ID to get work history for (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Work history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkHistoryEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
...
}