# src/routes/reputation-routes.ts
import { Router, Request, Response } from 'express';
import { authMiddleware } from '../middleware/auth-middleware.js';
import { validate, validateUUID, validateAppwriteDocumentId, submitRatingSchema } from '../middleware/validation-middleware.js';
import { apiRateLimiter } from '../middleware/rate-limiter.js';
import { getRequestId } from '../utils/route-helpers.js';
import { sendErrorResponse } from '../utils/response-helpers.js';
import { logger } from '../config/logger.js';
import {
  submitRating,
  getReputation,
  getWorkHistory,
  canUserRate,
} from '../services/reputation-service.js';
import {
  getAggregatedScore,
  getReputationBreakdown,
  getReputationHistory,
  getReputationLeaderboard,
} from '../services/reputation-aggregation-service.js';
import { asyncHandler } from '../utils/async-handler.js';

const router = Router();

/**
 * @swagger
 * components:
 *   schemas:
 *     BlockchainRating:
 *       type: object
 *       properties:
 *         id:
 *           type: string
 *         contractId:
 *           type: string
 *         raterId:
 *           type: string
 *         rateeId:
 *           type: string
 *         rating:
 *           type: integer
 *           minimum: 1
 *           maximum: 5
 *         comment:
 *           type: string
 *         timestamp:
 *           type: integer
 *         transactionHash:
 *           type: string
 *     ReputationScore:
 *       type: object
 *       properties:
 *         userId:
 *           type: string
 *         score:
 *           type: number
 *           description: Weighted average score with time decay
 *         totalRatings:
 *           type: integer
 *         averageRating:
 *           type: number
 *           description: Simple average without time decay
 *         ratings:
 *           type: array
 *           items:
 *             $ref: '#/components/schemas/BlockchainRating'
 *     WorkHistoryEntry:
 *       type: object
 *       properties:
 *         contractId:
 *           type: string
 *         projectId:
 *           type: string
 *         projectTitle:
 *           type: string
 *         role:
 *           type: string
 *           enum: [freelancer, employer]
 *         completedAt:
 *           type: string
 *           format: date-time
 *         rating:
 *           type: integer
 *           minimum: 1
 *           maximum: 5
 *         ratingComment:
 *           type: string
 *     RatingInput:
 *       type: object
 *       required:
 *         - contractId
 *         - rateeId
 *         - rating
 *       properties:
 *         contractId:
 *           type: string
 *         rateeId:
 *           type: string
 *         rating:
 *           type: integer
 *           minimum: 1
 *           maximum: 5
 *         comment:
 *           type: string
 */


// ============================================================
// IMPORTANT: Specific routes MUST be registered BEFORE parameterized routes
// Otherwise Express matches "can-rate" as a :userId parameter
// ============================================================

/**
 * @swagger
 * /api/reputation/can-rate:
 *   get:
 *     summary: Check if user can rate
 *     description: Check if the authenticated user can rate another user for a specific contract
 *     tags:
 *       - Reputation
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: query
 *         name: contractId
 *         required: true
 *         schema:
 *           type: string
 *         description: Contract ID
 *       - in: query
 *         name: rateeId
 *         required: true
 *         schema:
 *           type: string
 *         description: User ID to rate
 *     responses:
 *       200:
 *         description: Check completed successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 canRate:
 *                   type: boolean
 *                 reason:
 *                   type: string
 *       401:
 *         description: Unauthorized
 */
router.get('/can-rate', authMiddleware, apiRateLimiter, asyncHandler(async (req: Request, res: Response) => {
  const userId = req.user?.userId;
  const requestId = getRequestId(req);

  /* istanbul ignore next */
  if (!userId) {
    sendErrorResponse(res, 401, 'AUTH_UNAUTHORIZED', 'User not authenticated', { requestId });
    return;
  }

  const contractId = req.query['contractId'] as string | undefined;
  const rateeId = req.query['rateeId'] as string | undefined;

  if (!contractId || !rateeId) {
    sendErrorResponse(res, 400, 'VALIDATION_ERROR', 'contractId and rateeId are required query parameters', { requestId });
    return;
  }

  const result = await canUserRate(userId, rateeId, contractId);

  if (!result.success) {
    sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId });
    return;
  }

  res.status(200).json(result.data);
}));

/**
 * @swagger
 * /api/reputation/rate:
 *   post:
 *     summary: Submit rating
 *     description: Submit a rating for another user after contract completion. Rating must be between 1 and 5.
 *     tags:
 *       - Reputation
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/RatingInput'
 *     responses:
 *       201:
 *         description: Rating submitted successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 rating:
 *                   $ref: '#/components/schemas/BlockchainRating'
 *                 transactionHash:
 *                   type: string
 *       400:
 *         description: Invalid rating or validation error
 *       401:
 *         description: Unauthorized
 *       404:
 *         description: Contract not found
 *       409:
 *         description: Duplicate rating
 */
router.post('/rate', authMiddleware, apiRateLimiter, validate(submitRatingSchema), asyncHandler(async (req: Request, res: Response) => {
  const userId = req.user?.userId;
  const requestId = getRequestId(req);

  /* istanbul ignore next */
  if (!userId) {
    sendErrorResponse(res, 401, 'AUTH_UNAUTHORIZED', 'User not authenticated', { requestId });
    return;
  }

  // Required fields, UUID formats, and rating bounds are handled by the
  // middleware (submitRatingSchema); the service enforces the business rules.
  const { contractId, rateeId, rating, comment } = req.body as {
    contractId?: string;
    rateeId?: string;
    rating?: number;
    comment?: string;
  };

  const ratingInput: Parameters<typeof submitRating>[0] = {
    contractId: contractId!,
    raterId: userId,
    rateeId: rateeId!,
    rating: rating!,
  };
  if (comment) ratingInput.comment = comment;
  const result = await submitRating(ratingInput);

  if (!result.success) {
    let statusCode = 400;
    if (result.error.code === 'NOT_FOUND') statusCode = 404;
    if (result.error.code === 'UNAUTHORIZED') statusCode = 403;
    if (result.error.code === 'DUPLICATE_RATING') statusCode = 409;

    sendErrorResponse(res, statusCode, result.error.code, result.error.message, { requestId });
    return;
  }

  res.status(201).json(result.data);
}));


// ============================================================
// Specific routes MUST be registered BEFORE parameterized routes
// ============================================================

/**
 * @swagger
 * /api/reputation/leaderboard:
 *   get:
 *     summary: Get platform leaderboard
 *     tags:
 *       - Reputation
 *     parameters:
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *           default: 10
 *     responses:
 *       200:
 *         description: Top rated users
 */
router.get('/leaderboard', apiRateLimiter, asyncHandler(async (req: Request, res: Response) => {
  try {
    const limit = parseInt(req.query['limit'] as string) || 10;

    const result = await getReputationLeaderboard(limit);

    if (!result.success) {
      return sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId: getRequestId(req) });
    }

    return res.json(result.data);
  } catch (error) {
    logger.error('Error getting reputation leaderboard', { error });
    return sendErrorResponse(res, 500, 'INTERNAL_ERROR', 'Failed to get leaderboard', { requestId: getRequestId(req) });
  }
}));


// ============================================================
// Parameterized routes AFTER specific routes
// ============================================================

/**
 * @swagger
 * /api/reputation/{userId}:
 *   get:
 *     summary: Get user reputation
 *     description: Retrieves the reputation score and ratings for a user from the blockchain
 *     tags:
 *       - Reputation
 *     parameters:
 *       - in: path
 *         name: userId
 *         required: true
 *         schema:
 *           type: string
 *         description: User ID to get reputation for (Appwrite document ID)
 *     responses:
 *       200:
 *         description: Reputation retrieved successfully
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ReputationScore'
 *       400:
 *         description: Invalid UUID format
 *       404:
 *         description: User not found
 */
router.get('/:userId', apiRateLimiter, validateAppwriteDocumentId(['userId']), asyncHandler(async (req: Request, res: Response) => {
  const userId = req.params['userId'] ?? '';
  const requestId = getRequestId(req);

  /* istanbul ignore next */
  if (!userId) {
    sendErrorResponse(res, 400, 'VALIDATION_ERROR', 'User ID is required', { requestId });
    return;
  }

  const result = await getReputation(userId);

  if (!result.success) {
    sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId });
    return;
  }

  res.status(200).json(result.data);
}));

/**
 * @swagger
 * /api/reputation/{userId}/history:
 *   get:
 *     summary: Get work history
 *     description: Retrieves the work history for a user including completed contracts and ratings
 *     tags:
 *       - Reputation
 *     parameters:
 *       - in: path
 *         name: userId
 *         required: true
 *         schema:
 *           type: string
 *         description: User ID to get work history for (Appwrite document ID)
 *     responses:
 *       200:
 *         description: Work history retrieved successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: array
 *               items:
 *                 $ref: '#/components/schemas/WorkHistoryEntry'
 *       400:
 *         description: Invalid UUID format
 *       404:
 *         description: User not found
 */
router.get('/:userId/history', apiRateLimiter, validateAppwriteDocumentId(['userId']), asyncHandler(async (req: Request, res: Response) => {
  const userId = req.params['userId'] ?? '';
  const requestId = getRequestId(req);

  /* istanbul ignore next */
  if (!userId) {
    sendErrorResponse(res, 400, 'VALIDATION_ERROR', 'User ID is required', { requestId });
    return;
  }

  const result = await getWorkHistory(userId);

  if (!result.success) {
    sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId });
    return;
  }

  res.status(200).json(result.data);
}));

/**
 * @swagger
 * /api/reputation/{userId}/score:
 *   get:
 *     summary: Get aggregated reputation score
 *     tags:
 *       - Reputation
 *     parameters:
 *       - in: path
 *         name: userId
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Aggregated reputation score
 */
router.get('/:userId/score', validateAppwriteDocumentId(['userId']), apiRateLimiter, asyncHandler(async (req: Request, res: Response) => {
  try {
    const userId = req.params['userId'] ?? '';

    const result = await getAggregatedScore(userId);

    if (!result.success) {
      return sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId: getRequestId(req) });
    }

    return res.json(result.data);
  } catch (error) {
    logger.error('Error getting reputation score', { error });
    return sendErrorResponse(res, 500, 'INTERNAL_ERROR', 'Failed to get reputation score', { requestId: getRequestId(req) });
  }
}));

/**
 * @swagger
 * /api/reputation/{userId}/breakdown:
 *   get:
 *     summary: Get reputation breakdown
 *     tags:
 *       - Reputation
 *     parameters:
 *       - in: path
 *         name: userId
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Reputation breakdown by stars
 */
router.get('/:userId/breakdown', validateAppwriteDocumentId(['userId']), apiRateLimiter, asyncHandler(async (req: Request, res: Response) => {
  try {
    const userId = req.params['userId'] ?? '';

    const result = await getReputationBreakdown(userId);

    if (!result.success) {
      return sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId: getRequestId(req) });
    }

    return res.json(result.data);
  } catch (error) {
    logger.error('Error getting reputation breakdown', { error });
    return sendErrorResponse(res, 500, 'INTERNAL_ERROR', 'Failed to get reputation breakdown', { requestId: getRequestId(req) });
  }
}));

/**
 * @swagger
 * /api/reputation/{userId}/reputation-history:
 *   get:
 *     summary: Get reputation history over time
 *     tags:
 *       - Reputation
 *     parameters:
 *       - in: path
 *         name: userId
 *         required: true
 *         schema:
 *           type: string
 *       - in: query
 *         name: months
 *         schema:
 *           type: integer
 *           default: 12
 *     responses:
 *       200:
 *         description: Reputation history
 */
router.get('/:userId/reputation-history', validateAppwriteDocumentId(['userId']), apiRateLimiter, asyncHandler(async (req: Request, res: Response) => {
  try {
    const userId = req.params['userId'] ?? '';
    const months = parseInt(req.query['months'] as string) || 12;

    const result = await getReputationHistory(userId, months);

    if (!result.success) {
      return sendErrorResponse(res, 400, result.error.code, result.error.message, { requestId: getRequestId(req) });
    }

    return res.json(result.data);
  } catch (error) {
    logger.error('Error getting reputation history', { error });
    return sendErrorResponse(res, 500, 'INTERNAL_ERROR', 'Failed to get reputation history', { requestId: getRequestId(req) });
  }
}));

export default router;